+++ R/calculate_slope.R
@@ -4,6 +4,8 @@
 #'
 #' @param neighbours \code{numeric} value. Number of directions used in the Least Cost Path calculation. See Huber and Church (1985) for methodological considerations when choosing number of neighbours. Expected values are 4, 8, 16, 32, or 48. Default is 16
 #'
+#' @param exaggeration \code{logical}. if TRUE, positive slope values (ie. up-hill movement) multiplied by 1.99 and negative slope values (ie. down-hill movement) multiplied by 2.31.
+#'
 #' @noRd
 #'
 #' @import rgdal
@@ -15,15 +17,17 @@
 #'
 #' @author Joseph Lewis
 
-calculate_slope <- function(dem, neighbours) {
+calculate_slope <- function(dem, neighbours, exaggeration = exaggeration) {
     
-    altDiff_slope <- function(x) {
-        x[2] - x[1]
-    }
-    
-    hd <- suppressWarnings(gdistance::transition(x = dem, transitionFunction = altDiff_slope, directions = neighbours, symm = FALSE))
+    hd <- transition_slope(x = dem, neighbours)
     
     slope <- gdistance::geoCorrection(hd, scl = FALSE)
+    
+    if (exaggeration) {
+        
+        slope@transitionMatrix@x <- ifelse(slope@transitionMatrix@x > 0, slope@transitionMatrix@x * 1.99, slope@transitionMatrix@x * 2.31)
+        
+    }
     
     return(slope)